use thread local instead of pool for context
